+++ cardano-crypto-class/src/Cardano/Crypto/PackedBytes.hs
@@ -1,0 +1,11 @@
+module Cardano.Crypto.PackedBytes (
+  PackedBytes,
+  packByteString,
+  packShortByteString,
+  packShortByteStringWithOffset,
+  unpackBytes,
+  unpackAsByteArray,
+  unpackPinnedBytes,
+) where
+
+import Cardano.Crypto.PackedBytes.Internal